Move display to separate file
